+++ src/ext-lsv/lsv_cmd.cpp
@@ -1,0 +1,15 @@
+
+#include "ext-lsv/lsv_cmd.h"
+
+void init(Abc_Frame_t* pAbc) {
+  Cmd_CommandAdd(pAbc, "LSV", "lsv_print_nodes", Lsv_CommandPrintNodes, 0);
+  Cmd_CommandAdd(pAbc, "LSV", "lsv_print_sopunate", Lsv_CommandPrintSOPUnate, 0);
+}
+
+void destroy(Abc_Frame_t* pAbc) {}
+
+Abc_FrameInitializer_t frame_initializer = {init, destroy};
+
+struct PackageRegistrationManager {
+  PackageRegistrationManager() { Abc_FrameAddInitializer(&frame_initializer); }
+} lsvPackageRegistrationManager;
