+++ src/ext-lsv/lsv_print_nodes.cpp
@@ -1,21 +1,6 @@
-#include "base/abc/abc.h"
-#include "base/main/main.h"
-#include "base/main/mainInt.h"
 
-static int Lsv_CommandPrintNodes(Abc_Frame_t* pAbc, int argc, char** argv);
-
-void init(Abc_Frame_t* pAbc) {
-  Cmd_CommandAdd(pAbc, "LSV", "lsv_print_nodes", Lsv_CommandPrintNodes, 0);
-}
-
-void destroy(Abc_Frame_t* pAbc) {}
-
-Abc_FrameInitializer_t frame_initializer = {init, destroy};
-
-struct PackageRegistrationManager {
-  PackageRegistrationManager() { Abc_FrameAddInitializer(&frame_initializer); }
-} lsvPackageRegistrationManager;
-
+#include "ext-lsv/lsv_cmd.h"
+    
 void Lsv_NtkPrintNodes(Abc_Ntk_t* pNtk) {
   Abc_Obj_t* pObj;
   int i;
