+++ src/ext-lsv/lsv_print_sopunate.cpp
@@ -1,0 +1,8 @@
+
+#include "ext-lsv/lsv_cmd.h"
+
+int Lsv_CommandPrintSOPUnate(Abc_Frame_t* pAbc, int argc, char** argv)
+{
+    std::cout << "Hello Happy World" << std::endl;
+    return 0;
+}
