+++ src/ext-lsv/module.make
@@ -1,2 +1,4 @@
 SRC += \
-    src/ext-lsv/lsvCmd.cpp
+    src/ext-lsv/lsv_print_nodes.cpp \
+    src/ext-lsv/lsv_print_sopunate.cpp \
+    src/ext-lsv/lsv_cmd.cpp
